+++ src/View/MiniMapXYView.cpp
@@ -38,7 +38,7 @@ namespace TrenchBroom {
                               static_cast<float>(worldBounds.max.z()));
             
             m_camera->setNearPlane(0.0f);
-            m_camera->setFarPlane(worldBounds.size().z());
+            m_camera->setFarPlane(static_cast<float>(worldBounds.size().z()));
             m_camera->setDirection(Vec3f::NegZ, Vec3f::PosY);
             m_camera->moveTo(Vec3f(0.0f, 0.0f, worldBounds.max.z()));
             m_camera->setZoom(Vec2f(0.15f, 0.15f));
